requirements.txt
python-telegram-bot==13.15
SQLAlchemy==2.0.20
python-dotenv==1.0.0
pillow==10.0.0 